services/portfolio/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-dateutil==2.9.0.post0
